+++ Data/SBV/Dynamic.hs
@@ -128,17 +128,16 @@ import Data.SBV.Core.Concrete
 import Data.SBV.Core.Symbolic
 import Data.SBV.Core.Operations
 
-import Data.SBV.Compilers.CodeGen
-  ( SBVCodeGen
-  , svCgInput, svCgInputArr
-  , svCgOutput, svCgOutputArr
-  , svCgReturn, svCgReturnArr
-  , cgPerformRTCs, cgSetDriverValues, cgGenerateDriver, cgGenerateMakefile
-  , cgAddPrototype, cgAddDecl, cgAddLDFlags, cgIgnoreSAssert
-  , cgIntegerSize, cgSRealType, CgSRealType(..)
-  )
-import Data.SBV.Compilers.C    (compileToC, compileToCLib)
+import Data.SBV.Compilers.CodeGen ( SBVCodeGen
+                                  , svCgInput, svCgInputArr
+                                  , svCgOutput, svCgOutputArr
+                                  , svCgReturn, svCgReturnArr
+                                  , cgPerformRTCs, cgSetDriverValues, cgGenerateDriver, cgGenerateMakefile
+                                  , cgAddPrototype, cgAddDecl, cgAddLDFlags, cgIgnoreSAssert
+                                  , cgIntegerSize, cgSRealType, CgSRealType(..)
+                                  )
 import Data.SBV.Compilers.C       (compileToC, compileToCLib)
+
 import Data.SBV.Provers.Prover (boolector, cvc4, yices, z3, mathSAT, abc, defaultSMTCfg)
 import Data.SBV.SMT.SMT        (ThmResult(..), SatResult(..), SafeResult(..), OptimizeResult(..), AllSatResult(..), genParse)
 import Data.SBV                (sbvCurrentSolver, sbvCheckSolverInstallation, defaultSolverConfig, sbvAvailableSolvers)