+++ Data/SBV/Examples/BitPrecise/MergeSort.hs
@@ -12,6 +12,7 @@
 module Data.SBV.Examples.BitPrecise.MergeSort where
 
 import Data.SBV
+import Data.SBV.Tools.CodeGen
 
 -----------------------------------------------------------------------------
 -- * Implementing Merge-Sort