+++ Data/SBV/Examples/Crypto/AES.hs
@@ -29,7 +29,9 @@
 module Data.SBV.Examples.Crypto.AES where
 
 import Data.SBV
+import Data.SBV.Tools.CodeGen
 import Data.SBV.Tools.Polynomial
+
 import Data.List (transpose)
 
 -----------------------------------------------------------------------------