+++ SBVUnitTest/TestSuite/BitPrecise/MergeSort.hs
@@ -11,7 +11,6 @@
 
 module TestSuite.BitPrecise.MergeSort where
 
-import Data.SBV
 import Data.SBV.Internals
 import Data.SBV.Examples.BitPrecise.MergeSort