+++ SBVUnitTest/TestSuite/CodeGeneration/Uninterpreted.hs
@@ -11,7 +11,6 @@
 
 module TestSuite.CodeGeneration.Uninterpreted(testSuite) where
 
-import Data.SBV
 import Data.SBV.Internals
 import Data.SBV.Examples.CodeGeneration.Uninterpreted